# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  PENDING: 'PENDING',
} as const;

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  USER: 'USER',
  OWNER: 'OWNER',
  STAFF: 'STAFF',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  username: string;
  status: UserStatus;
  role: UserRole;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  lastLoginAt: string | null;
}

export interface AuthResponse {
  user: User;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  email: string;
  /**
     * @minLength 3
     * @maxLength 40
     * @pattern ^[a-zA-Z0-9_]+$
     */
  username: string;
  /**
     * @minLength 10
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}

export type WorldStatus = typeof WorldStatus[keyof typeof WorldStatus];


export const WorldStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  SUSPENDED: 'SUSPENDED',
} as const;

export interface World {
  id: string;
  ownerId: string;
  name: string;
  slug: string;
  status: WorldStatus;
  createdAt: string;
  updatedAt: string;
}

export interface WorldInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 3
     * @maxLength 80
     * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
     */
  slug: string;
}

export type AccessCheckRole = typeof AccessCheckRole[keyof typeof AccessCheckRole];


export const AccessCheckRole = {
  USER: 'USER',
  OWNER: 'OWNER',
  STAFF: 'STAFF',
} as const;

export interface AccessCheck {
  allowed: boolean;
  permission: string;
  role: AccessCheckRole;
}

export type CheckAccessParams = {
/**
 * @minLength 1
 * @maxLength 120
 */
permission: string;
};