# lib/api-zod/src/generated/types/accessCheck.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccessCheckRole } from './accessCheckRole';

export interface AccessCheck {
  allowed: boolean;
  permission: string;
  role: AccessCheckRole;
}