# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  email: string;
  /**
     * @minLength 3
     * @maxLength 40
     * @pattern ^[a-zA-Z0-9_]+$
     */
  username: string;
  /**
     * @minLength 10
     * @maxLength 128
     */
  password: string;
}