# lib/api-zod/src/generated/types/world.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorldStatus } from './worldStatus';

export interface World {
  id: string;
  ownerId: string;
  name: string;
  slug: string;
  status: WorldStatus;
  createdAt: Date;
  updatedAt: Date;
}